Use only product attributes when adding to cart
Previously, when adding a variation product to cart using
`add_to_cart()` we were sending not only the product attributes,
but all the form data sent to `wc_ajax_generate_cart` via $_POST.

Now, all of the attributes are added to a property called `attributes`
in the ajax call. And, when adding the variation product to the cart
only these attributes are used.