Reference Txns: Send address only if complete
If all of the required address fields are not sent,
PayPal responds with an error about the missing address fields:
- name
- street 1
- city
- country
- zipcode (if the country is US)
- state (for some countries)

To fix this, send the address parameters only if all
the required fields are available.

Note that this fix is in the function
`get_do_reference_transaction_params()`, and it's currently only
being used for subscription renewals.

This was affecting virtual subscriptions because
their shipping address was not always saved.