Map languages for US based stores
For stores based in the US we need to map the language code to a xx_US based code to ensure
PayPal Credit is displayed. PayPal supports en_US, fr_US, es_US, and zh_US. For stores based in the US,
we will always map it to _US based locale codes.