Simplify error handling
It is not possible to “silently” cancel a promise while inside a `createOrder()` callback. The `onError()` handler is always called when throwing an error, so it becomes impossible to differentiate between genuine errors we can expect from `createOrder()` or generic errors thrown elsewhere.

This changes makes `onError()` do nothing for generic errors, so that errors explicitly displayed via `showError()` are not overridden.