Move js enqueuing to display_paypal_button_product
Rather than checking for the relevant form element in single product
page and returning early in wc-gateway-ppec-generate-cart.js, this
instead moves the enqueuing of the relevant script to
display_paypal_button_product() in class-wc-gateway-ppec-cart-handler.php

One of the conditions of the above display function is whether a product
is in stock. So rather than run the script at all in cases where a
product is not in stock, moving the enqueue function here ensures the
script is not enqueued at all in these cases.