Allow WC to perform checks when missing fields
Previously, when purchasing variable products, or products with required
fields (Product Add-ons), the PPEC buttons did not trigger the same
behavior as the default Add To Cart WC button. This led to some
disconnect between the two options.

This adds an event listener to the PPEC button container and forces the
Add To Cart button click when all required fields have not been
submitted. This in turn enforces the default WC behavior for these
cases, and gets rid of that disconnect.